+++ BLL/Representation/VehiculeRepresentation.cs
@@ -13,10 +13,10 @@ namespace BLL.Representation
         public string photoRepresentation { get; set; }
         public string marqueRepresentation { get; set; }
         public string typeMoteurRepresentation { get; set; }
-        public double prixRepresentation { get; set; }
+        public decimal prixRepresentation { get; set; }
         public string estDispoRepresentation { get; set; }
         public string promoRepresentation { get; set; } = string.Empty;
-        public double promoPrixRepresentation { get; set; }
+        public decimal promoPrixRepresentation { get; set; }
 
     }
 }
